fix: migrate yoga-api to Prisma 7 runtime requirements
Prisma 7 removed the legacy binary/library engine and now requires a
driver adapter for direct DB connections. Key changes:

- Install @prisma/adapter-pg + pg for direct PostgreSQL connection
- Update src/db/client.ts to use PrismaPg adapter
- Remove url from schema.prisma datasource (Prisma 7 no longer supports it)
- Add generateDatamodel=true to Pothos generator and change output to .ts
  so getDatamodel() has a runtime implementation (not just type declaration)
- Update builder.ts to import getDatamodel() from generated types instead
  of passing Prisma.dmmf (which no longer includes uniqueIndexes in v7)
- Keep prisma.config.ts for CLI commands (migrate, generate, studio)
