+++ yoga-api/package.json
@@ -22,16 +22,19 @@
   "dependencies": {
     "@pothos/core": "4.12.0",
     "@pothos/plugin-prisma": "4.14.2",
+    "@prisma/adapter-pg": "7.4.2",
     "@prisma/client": "7.4.2",
     "busboy": "1.6.0",
     "dotenv": "16.5.0",
     "graphql": "16.13.1",
     "graphql-yoga": "5.18.1",
+    "pg": "8.20.0",
     "zod": "4.3.6"
   },
   "devDependencies": {
     "@types/busboy": "1.5.4",
     "@types/node": "25.4.0",
+    "@types/pg": "8.11.14",
     "@typescript-eslint/eslint-plugin": "8.57.0",
     "@typescript-eslint/parser": "8.57.0",
     "eslint": "10.0.3",
